Save sign-up status in auth component
Update activation status saved in Auth component so that the
same information is displayed if the user switches back and
forth from the sign-up to the sign-in tabs, regardless of
user's stage in the sign up process. Previously, the sign-up tab
would be reset if the user switched to the sign-in tab and then
back again to the sign-up tab.